encapsulate ._decoded_secrets + document structs
